tests/wpt/tests/css/css-animations/parsing/animation-iteration-count-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-iteration-count with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-iteration-count">
<meta name="assert" content="animation-iteration-count supports the full grammar '<single-animation-iteration-count> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-iteration-count", "0");
test_valid_value("animation-iteration-count", "3");
test_valid_value("animation-iteration-count", "4.5");
test_valid_value("animation-iteration-count", "infinite");

test_valid_value("animation-iteration-count", "0, infinite, 3");
</script>
</body>
</html>
